pyproject.toml
[tool.poetry]
name = "contacts-proj"
version = "0.1.0"
description = ""
authors = ["dima cheban <dimacheban23@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
sqlalchemy = "^2.0.36"
pydantic-settings = "^2.7.0"
pydantic = {extras = ["email"], version = "^2.10.4"}
asyncpg = "^0.30.0"
python-dotenv = "^1.0.1"
greenlet = "^3.1.1"
alembic = "^1.14.0"
redis = "^5.2.1"
jwt = "^1.3.1"
pytz = "^2024.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = "^1.7.4"
aio-pika = "^9.5.4"
fastapi-mail = "^1.4.2"
libgravatar = "^1.0.4"
bcrypt = "^4.2.1"
python-multipart = "^0.0.20"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

packeges = [
    { include = 'app' }
]